+++ src/modules/PlayerBots/playerbot/strategy/actions/ActionContext.h
@@ -458,6 +458,10 @@ namespace ai
             creators["engage thekal zealot"] = [](PlayerbotAI* ai) { return new EngageThekalZealotAction(ai); };
             creators["engage thekal tiger"] = [](PlayerbotAI* ai) { return new EngageThekalTigerAction(ai); };
             creators["engage arlokk prowler"] = [](PlayerbotAI* ai) { return new EngageArlokkProwlerAction(ai); };
+            creators["engage venoxis cobra"] = [](PlayerbotAI* ai) { return new EngageVenoxisCobraAction(ai); };
+            creators["engage gothik add"] = [](PlayerbotAI* ai) { return new EngageGothikAddAction(ai); };
+            creators["engage yauj brood"] = [](PlayerbotAI* ai) { return new EngageYaujBroodAction(ai); };
+            creators["engage cthun tentacle"] = [](PlayerbotAI* ai) { return new EngageCthunTentacleAction(ai); };
 
             // AQ20 dungeon strategy enable/disable
             creators["enable ruins of ahnqiraj strategy"] = [](PlayerbotAI* ai) { return new RuinsOfAhnQirajEnableDungeonStrategyAction(ai); };